Reconfigured files to use router
