+++ src/pages/About.vue
@@ -1,0 +1,9 @@
+<template>
+  <p>About CourseRoad</p>
+</template>
+
+<script>
+export default {
+  name: "About"
+}
+</script>
